Fix cell vertical alignment trouble

+++ src/com/lowagie/text/pdf/PdfText.java
@@ -4,7 +4,7 @@
  *               rugPdf0.20:		0.14 99/12/01
  *               iText0.3:			0.24 2000/02/14
  *               iText0.35:         0.24 2000/08/11
- * 
+ *
  * Copyright (c) 1999, 2000 Bruno Lowagie.
  *
  * This library is free software; you can redistribute it and/or modify it
@@ -31,7 +31,7 @@
  * BELGIUM
  * tel. +32 (0)9 228.10.97
  * bruno@lowagie.com
- *  
+ *
  */
 
 package com.lowagie.text.pdf;
@@ -101,29 +101,29 @@ class PdfText extends PdfContent {
 // methods concerning the membervariables
 
 	/**
-     * Returns the PDF representation of this <CODE>PdfText</CODE>-object.
+		 * Returns the PDF representation of this <CODE>PdfText</CODE>-object.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.2 (page 344-345).
 	 *
 	 * @return		a <CODE>String</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final String toPdf() {
+		final String toPdf() {
 		String text = toString();
 		if (text.length() == 0) {
 			return "";
 		}
 		return new StringBuffer("BT\n").append(text).append("ET").toString();
-    }
+		}
 
 	/**
 	 * Returns the startposition of the <VAR>current line</VAR>.
 	 *
 	 * @return		a coordinate
-     *
+		 *
 	 * @since		iText0.30
 	 */
 
@@ -135,7 +135,7 @@ class PdfText extends PdfContent {
 	 * Returns the y-coordinate of the <VAR>current point</VAR>.
 	 *
 	 * @return		a coordinate
-     *
+		 *
 	 * @since		rugPdf0.10
 	 */
 
@@ -147,113 +147,113 @@ class PdfText extends PdfContent {
 	 * Returns the leading.
 	 *
 	 * @return		the leading value
-     *
+		 *
 	 * @since		rugPdf0.10
 	 */
 
 	final int leading() {
 		return leading;
-	} 
+	}
 
-// methods to add textual content 
+// methods to add textual content
 
 	/**
-     * Adds the content of another <CODE>PdfText</CODE>-object to this object.
+		 * Adds the content of another <CODE>PdfText</CODE>-object to this object.
 	 *
 	 * @param		other
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void add(PdfText other) {
+		final void add(PdfText other) {
 		offset += other.offset();
 		y += other.y();
 		content.append(other.toString());
-    }
+		}
 
 	/**
-     * Sets the character spacing parameter.
+		 * Sets the character spacing parameter.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.1.1 (page 340).</P>
 	 *
 	 * @param		charSpace			a parameter
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setCharacterSpacing(int charSpace) {
+		final void setCharacterSpacing(int charSpace) {
 		content.append(charSpace).append(" Tc ");
-    } 
+		}
 
 	/**
-     * Sets the character spacing parameter.
+		 * Sets the character spacing parameter.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.1.1 (page 340).</P>
 	 *
 	 * @param		charSpace			a parameter
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setCharacterSpacing(double charSpace) {
+		final void setCharacterSpacing(double charSpace) {
 		content.append(charSpace).append(" Tc ");
-    } 
+		}
 
 	/**
-     * Sets the word spacing parameter.
+		 * Sets the word spacing parameter.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.1.2 (page 340-341).</P>
 	 *
 	 * @param		wordSpace			a parameter
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setWordSpacing(int wordSpace) {
+		final void setWordSpacing(int wordSpace) {
 		content.append(wordSpace).append(" Tw ");
-    } 
+		}
 
 	/**
-     * Sets the word spacing parameter.
+		 * Sets the word spacing parameter.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.1.2 (page 340-341).</P>
 	 *
 	 * @param		wordSpace			a parameter
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setWordSpacing(double wordSpace) {
+		final void setWordSpacing(double wordSpace) {
 		content.append(wordSpace).append(" Tw ");
-    } 														  
+		}
 
 	/**
-     * Sets the horizontal scaling parameter.
+		 * Sets the horizontal scaling parameter.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.1.3 (page 341-342).</P>
 	 *
 	 * @param		scale				a parameter
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setHorizontalScaling(int scale) {
+		final void setHorizontalScaling(int scale) {
 		content.append(scale).append(" Tz\n");
-    } 
+		}
 
 	/**
-     * Sets the text leading parameter.
+		 * Sets the text leading parameter.
 	 * <P>
 	 * The leading parameter is measured in text space units. It specifies the vertical distance
 	 * between the baselines of adjacent lines of text.</P>
@@ -263,49 +263,49 @@ class PdfText extends PdfContent {
 	 *
 	 * @param		leading			the new leading
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setLeading(int leading) {
+		final void setLeading(int leading) {
 		this.leading = leading;
 		content.append(leading).append(" TL\n");
-    }
+		}
 
 	/**
-     * Changes the font.
+		 * Changes the font.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.1.5 (page 342).</P>
 	 *
 	 * @param		font		a <CODE>PdfFont</CODE>-object
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setFont(PdfFont font) {
+		final void setFont(PdfFont font) {
 		content.append(font.getName().toString()).append(' ').append(font.size()).append(" Tf\n");
-    }													  
+		}
 
 	/**
-     * Sets the text rendering parameter.
+		 * Sets the text rendering parameter.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.1.7 (page 343-344).</P>
 	 *
 	 * @param		rendering				a parameter
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setRenderingMode(int rendering) {
+		final void setRenderingMode(int rendering) {
 		content.append(rendering).append(" Tr ");
-    }														   
+		}
 
 	/**
-     * Sets the text rise parameter.
+		 * Sets the text rise parameter.
 	 * <P>
 	 * This allows to write text in subscript or superscript mode.</P>
 	 * <P>
@@ -314,16 +314,16 @@ class PdfText extends PdfContent {
 	 *
 	 * @param		rise				a parameter
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setTextRise(int rise) {
+		final void setTextRise(int rise) {
 		content.append(rise).append(" Ts ");
-    } 
+		}
 
 	/**
-     * Moves to the start of the next line, offset from the start of the current line.
+		 * Moves to the start of the next line, offset from the start of the current line.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.3 (page 345-346).</P>
@@ -331,18 +331,18 @@ class PdfText extends PdfContent {
 	 * @param		x			x-coordinate of the new current point
 	 * @param		y			y-coordinate of the new current point
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void move(int x, int y) {
+		final void move(float x, float y) {
 		this.offset += x;
 		this.y += y;
 		content.append(x).append(' ').append(y).append(" Td\n");
-    } 
+		}
 
 	/**
-     * Moves to the start of the next line, offset from the start of the current line.
+		 * Moves to the start of the next line, offset from the start of the current line.
 	 * <P>
 	 * As a side effect, this sets the leading parameter in the text state.</P>
 	 * <P>
@@ -352,19 +352,19 @@ class PdfText extends PdfContent {
 	 * @param		x			offset of the new current point
 	 * @param		y			y-coordinate of the new current point
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void moveWithLeading(int x, int y) {
+		final void moveWithLeading(int x, int y) {
 		this.offset += x;
 		this.y += y;
 		leading = -y;
 		content.append(x).append(' ').append(y).append(" TD\n");
-    }
+		}
 
 	/**
-     * Changes the text matrix.
+		 * Changes the text matrix.
 	 * <P>
 	 * Remark: this operation also initializes the current point position.</P>
 	 * <P>
@@ -376,100 +376,100 @@ class PdfText extends PdfContent {
 	 * @param		c			operand 2,1 in the matrix
 	 * @param		d			operand 2,2 in the matrix
 	 * @param		x			operand 3,1 in the matrix
-	 * @param		y			operand 3,2 in the matrix			
+	 * @param		y			operand 3,2 in the matrix
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void setMatrix(int a, int b, int c, int d, int x, int y) {
+		final void setMatrix(int a, int b, int c, int d, int x, int y) {
 		content.append(a).append(' ').append(b).append(' ').append(c).append(' ').append(d).append(' ').append(x).append(' ').append(y).append(" Tm\n");
-    } 
+		}
 
 	/**
-     * Moves to the start of the next line.
+		 * Moves to the start of the next line.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.3 (page 345-346).</P>
 	 *
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void newLine() {
+		final void newLine() {
 		this.y -= leading;
 		content.append("T* ");
-    }
+		}
 
 	/**
-     * Shows text string, using the character and word spacing parameters from the text state.
+		 * Shows text string, using the character and word spacing parameters from the text state.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.5 (page 346-347).
 	 *
 	 * @param		text			a <CODE>PdfPrintable</CODE>
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void show(PdfPrintable text) {
+		final void show(PdfPrintable text) {
 		content.append(new PdfString(text).get()).append(" Tj\n");
-    } 
+		}
 
 	/**
-     * Shows text string, using the character and word spacing parameters from the text state.
+		 * Shows text string, using the character and word spacing parameters from the text state.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.5 (page 346-347).
 	 *
 	 * @param		text			a <CODE>String</CODE>
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void show(String text) {
+		final void show(String text) {
 		content.append(new PdfString(text).get()).append(" Tj\n");
-    } 
+		}
 
 	/**
-     * Moves to the next line and shows text string, using the character and word spacing parameters from the text state.
+		 * Moves to the next line and shows text string, using the character and word spacing parameters from the text state.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.5 (page 346-347).
 	 *
 	 * @param		text			a <CODE>PdfPrintable</CODE>
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void newLineShow(PdfPrintable text) {
+		final void newLineShow(PdfPrintable text) {
 		this.y -= leading;
 		content.append(new PdfString(text).get()).append(" '\n");
-    } 
+		}
 
 	/**
-     * Moves to the next line and shows text string, using the character and word spacing parameters from the text state.
+		 * Moves to the next line and shows text string, using the character and word spacing parameters from the text state.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.5 (page 346-347).
 	 *
 	 * @param		text			a <CODE>String</CODE>
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
 	 */
 
-    final void newLineShow(String text) {
+		final void newLineShow(String text) {
 		this.y -= leading;
 		content.append(new PdfString(text).get()).append(" '\n");
-    } 
+		}
 
 	/**
-     * Moves to the next line and shows text string, using the given values of the character and word spacing parameters.
+		 * Moves to the next line and shows text string, using the given values of the character and word spacing parameters.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.5 (page 346-347).
@@ -478,17 +478,17 @@ class PdfText extends PdfContent {
 	 * @param		charSpacing		a parameter
 	 * @param		text			a <CODE>PdfPrintable</CODE>
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void newLineShow(int wordSpacing, int charSpacing, PdfPrintable text) {
+		final void newLineShow(int wordSpacing, int charSpacing, PdfPrintable text) {
 		this.y -= leading;
 		content.append(wordSpacing).append(' ').append(charSpacing).append(' ').append(new PdfString(text).get()).append(" \"\n");
-    } 
+		}
 
 	/**
-     * Moves to the next line and shows text string, using the given values of the character and word spacing parameters.
+		 * Moves to the next line and shows text string, using the given values of the character and word spacing parameters.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.5 (page 346-347).
@@ -497,28 +497,28 @@ class PdfText extends PdfContent {
 	 * @param		charSpacing		a parameter
 	 * @param		text			a <CODE>String</CODE>
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void newLineShow(int wordSpacing, int charSpacing, String text) {
+		final void newLineShow(int wordSpacing, int charSpacing, String text) {
 		this.y -= leading;
 		content.append(wordSpacing).append(' ').append(charSpacing).append(' ').append(new PdfString(text).get()).append(" \"\n");
-    }
+		}
 
 	/**
-     * Show with displacements.
+		 * Show with displacements.
 	 * <P>
 	 * This method is described in the 'Portable Document Format Reference Manual version 1.3'
 	 * section 8.7.5 (page 346-347).
 	 *
 	 * @param		text			a <CODE>PdfArray</CODE>
 	 * @return		<CODE>void</CODE>
-     *
+		 *
 	 * @since		rugPdf0.10
-     */
+		 */
 
-    final void show(PdfTextArray text) {
+		final void show(PdfTextArray text) {
 		content.append(text.toString()).append(" TJ\n");
-    }
+		}
 }
